Moved a couple of services from `testingNative` down to `platformNative`
