+++ subprojects/platform-native/src/main/java/org/gradle/nativeplatform/toolchain/internal/xcode/MacOSSdkPlatformPathLocator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package org.gradle.nativeplatform.test.xctest.internal;
+package org.gradle.nativeplatform.toolchain.internal.xcode;
 
-import org.gradle.language.swift.internal.AbstractLocator;
 import org.gradle.process.internal.ExecActionFactory;
 
 import javax.inject.Inject;
