--- subprojects/testing-native/src/main/java/org/gradle/nativeplatform/test/xctest/internal/services/XCTestTestingServices.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.nativeplatform.test.xctest.internal.services;
-
-import org.gradle.internal.service.ServiceRegistration;
-import org.gradle.internal.service.scopes.AbstractPluginServiceRegistry;
-import org.gradle.nativeplatform.test.xctest.internal.MacOSSdkPlatformPathLocator;
-
-public class XCTestTestingServices extends AbstractPluginServiceRegistry {
-    @Override
-    public void registerBuildSessionServices(ServiceRegistration registration) {
-        registration.add(MacOSSdkPlatformPathLocator.class);
-    }
-}
