+++ subprojects/testing-native/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.PluginServiceRegistry
@@ -1,2 +1,1 @@
 org.gradle.nativeplatform.test.internal.services.NativeTestingServices
-org.gradle.nativeplatform.test.xctest.internal.services.XCTestTestingServices
